Move TypeVisitor into its own file
